fix(PR-57): address review comments for US-0403
- Remove 6 duplicate service files from Infrastructure/Services
  (AddressService, AuthService, CartService, CategoryService,
  OrderService, ProductService) — canonical versions live in
  Application/Services
- Move ReviewService to Application/Services and refactor to use
  IGenericRepository<T> pattern instead of direct AppDbContext
- Remove UpdateReview endpoint + UpdateReviewDto (US-0404 scope)
- Revert AddressController.cs (pincode endpoint out of scope,
  also had new HttpClient() memory leak)
- Revert ECommerce.Application.csproj whitespace-only formatting
- Clean up ReviewController indentation and DTOs formatting
