+++ Backend/ECommerce.API/Controllers/ReviewController.cs
@@ -1,4 +1,4 @@
-﻿using System.Security.Claims;
+using System.Security.Claims;
 using ECommerce.Application.DTOs;
 using ECommerce.Application.Interfaces;
 using Microsoft.AspNetCore.Authorization;
@@ -21,19 +21,16 @@ namespace ECommerce.API.Controllers
         [Authorize]
         public async Task<IActionResult> CreateReview([FromBody] CreateReviewDto dto)
         {
-
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
             try
             {
-                // ✅ Extract UserId from JWT
                 var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
-
                 if (userIdClaim == null)
                     return Unauthorized("Invalid token");
 
                 var userId = int.Parse(userIdClaim.Value);
-
                 var result = await _reviewService.CreateReviewAsync(dto, userId);
 
                 return Ok(new
@@ -44,58 +41,19 @@ namespace ECommerce.API.Controllers
             }
             catch (KeyNotFoundException ex)
             {
-                return NotFound(ex.Message); // 404
-            }
-            catch (ArgumentException ex)
-            {
-                return BadRequest(ex.Message); // invalid rating/comment
-            }
-            catch (InvalidOperationException ex)
-            {
-                return BadRequest(ex.Message); // duplicate review
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, new
-                {
-                    message = "Internal server error"
-                });
-            }
-        }
-    
-
-    [HttpPut("{reviewId}")]
-        [Authorize]
-        public async Task<IActionResult> UpdateReview(int reviewId, [FromBody] UpdateReviewDto dto)
-        {
-            try
-            {
-                var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
-
-                if (userIdClaim == null)
-                    return Unauthorized("Invalid token");
-
-                var userId = int.Parse(userIdClaim.Value);
-
-                var result = await _reviewService.UpdateReviewAsync(reviewId, dto, userId);
-
-                return Ok(new
-                {
-                    message = "Review updated successfully",
-                    data = result
-                });
-            }
-            catch (KeyNotFoundException ex)
-            {
                 return NotFound(ex.Message);
-            }
-            catch (UnauthorizedAccessException ex)
-            {
-                return Unauthorized(ex.Message);
             }
             catch (ArgumentException ex)
             {
                 return BadRequest(ex.Message);
+            }
+            catch (InvalidOperationException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+            catch (Exception)
+            {
+                return StatusCode(500, new { message = "Internal server error" });
             }
         }
 
@@ -126,6 +84,5 @@ namespace ECommerce.API.Controllers
                 return NotFound(ex.Message);
             }
         }
-     }
+    }
 }
-
