+++ Backend/ECommerce.Application/DTOs/CreateReviewDto.cs
@@ -1,10 +1,9 @@
-﻿namespace ECommerce.Application.DTOs
+namespace ECommerce.Application.DTOs
 {
-      public class CreateReviewDto
-        {
-            public int ProductId { get; set; }
-            public int Rating { get; set; }
-            public string Comment { get; set; }
-        }
-    
+    public class CreateReviewDto
+    {
+        public int ProductId { get; set; }
+        public int Rating { get; set; }
+        public string Comment { get; set; } = string.Empty;
+    }
 }
