+++ Backend/ECommerce.Application/DTOs/ProductRatingDto.cs
@@ -1,9 +1,8 @@
-﻿namespace ECommerce.Application.DTOs
+namespace ECommerce.Application.DTOs
 {
-        public class ProductRatingDto
-        {
-            public double AverageRating { get; set; }
-            public int TotalReviews { get; set; }
-        }
-    
+    public class ProductRatingDto
+    {
+        public double AverageRating { get; set; }
+        public int TotalReviews { get; set; }
+    }
 }
