+++ Backend/ECommerce.Application/DTOs/ReviewDto.cs
@@ -1,13 +1,13 @@
-﻿namespace ECommerce.Application.DTOs
+namespace ECommerce.Application.DTOs
 {
     public class ReviewDto
     {
         public int Id { get; set; }
         public int ProductId { get; set; }
         public int UserId { get; set; }
-        public string UserName { get; set; }
+        public string UserName { get; set; } = string.Empty;
         public int Rating { get; set; }
-        public string Comment { get; set; }
+        public string Comment { get; set; } = string.Empty;
         public DateTime CreatedAt { get; set; }
     }
 }
