--- Backend/ECommerce.Application/DTOs/UpdateReviewDto.cs
@@ -1,8 +1,0 @@
-﻿namespace ECommerce.Application.DTOs
-{
-    public class UpdateReviewDto
-    {
-        public int Rating { get; set; }
-        public string Comment { get; set; }
-    }
-}
