+++ Backend/ECommerce.Application/Interfaces/IReviewService.cs
@@ -1,11 +1,10 @@
-﻿using ECommerce.Application.DTOs;
+using ECommerce.Application.DTOs;
 
 namespace ECommerce.Application.Interfaces
 {
     public interface IReviewService
     {
         Task<ReviewDto> CreateReviewAsync(CreateReviewDto dto, int userId);
-        Task<ReviewDto> UpdateReviewAsync(int reviewId, UpdateReviewDto dto, int userId);
         Task<List<ReviewDto>> GetReviewsByProductIdAsync(int productId);
         Task<ProductRatingDto> GetProductRatingAsync(int productId);
     }
