+++ Backend/ECommerce.Infrastructure/DependencyInjection.cs
@@ -3,7 +3,6 @@ using ECommerce.Application.Services;
 using ECommerce.Core.Interfaces;
 using ECommerce.Infrastructure.Data;
 using ECommerce.Infrastructure.Repositories;
-using ECommerce.Infrastructure.Services;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
